pyproject.toml
[tool.poetry]
name = "hx_export"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "^1.4.2"
base-repr = "^1.0.5"
google-cloud-bigquery = "^3.1.0"
python-dotenv = "^0.20.0"
pandas-gbq = "^0.17.5"
db-dtypes = "^1.0.1"
tqdm = "^4.64.0"
tomli = "^2.0.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
